Refactor database helpers and routes
