+++ magazyn/auth.py
@@ -1,0 +1,12 @@
+from functools import wraps
+from flask import session, redirect, url_for
+
+
+def login_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if 'username' not in session:
+            return redirect(url_for('login'))
+        return f(*args, **kwargs)
+
+    return decorated_function
