+++ magazyn/templates/add_item.html
@@ -2,7 +2,7 @@
 
 {% block content %}
 <h2>Dodaj nowy przedmiot</h2>
-    <form action="{{ url_for('add_item') }}" method="post">
+    <form action="{{ url_for('products.add_item') }}" method="post">
         <label for="name">Nazwa produktu:</label>
         <input type="text" id="name" name="name" required>
