+++ magazyn/templates/base.html
@@ -15,9 +15,9 @@
         <nav>
             <ul>
                 <li><a href="{{ url_for('home') }}">Strona główna</a></li>
-                <li><a href="{{ url_for('add_item') }}">Dodaj przedmiot</a></li>
-                <li><a href="{{ url_for('items') }}">Przedmioty</a></li>
-                <li><a href="{{ url_for('print_history') }}">Historia drukowania</a></li>
+                <li><a href="{{ url_for('products.add_item') }}">Dodaj przedmiot</a></li>
+                <li><a href="{{ url_for('products.items') }}">Przedmioty</a></li>
+                <li><a href="{{ url_for('history.print_history') }}">Historia drukowania</a></li>
                 <li><a href="{{ url_for('agent_logs') }}">Logi</a></li>
                 <li><a href="{{ url_for('settings') }}">Ustawienia</a></li>
                 <li><a href="{{ url_for('test_print') }}">Testuj drukarkę</a></li>
