+++ magazyn/templates/import_products.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <h2>Import produktów z pliku Excel</h2>
-    <form action="{{ url_for('import_products') }}" method="POST" enctype="multipart/form-data">
+    <form action="{{ url_for('products.import_products') }}" method="POST" enctype="multipart/form-data">
         <input type="file" name="file" required>
         <button type="submit">Zaimportuj</button>
     </form>
